+++ src/friends/module-loader/index.ts
@@ -1,0 +1,17 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+/**
+ * [[include:friends/module-loader/README.md]]
+ * @packageDocumentation
+ */
+
+export { default } from 'friends/module-loader/class';
+
+export * from 'friends/module-loader/api';
+export * from 'friends/module-loader/interface';